Extract serialize_response function to a new module
Moved the serialize_response function from address.py to a new module, serialize_response.py, for better modularity and reusability across different modules. This change enhances code organization by separating concerns and makes the serialize_response function independently accessible.
