+++ pan_scm_ansible_collection/plugins/module_utils/serialize_response.py
@@ -1,0 +1,9 @@
+def serialize_response(response):
+    """Convert response to Ansible-compatible format."""
+    if hasattr(response, 'model_dump'):
+        data = response.model_dump()
+        # Convert UUID to string
+        if 'id' in data and data['id']:
+            data['id'] = str(data['id'])
+        return data
+    return response
